fix: resolve build errors in ingredient search backend
Fix CA1805 warning (explicit default value) in IngredientSearchRequestDto,
cast TagCategory enum to int for RecipeTagDto mapping, and upgrade Npgsql
to 9.0.3 to match EF Core dependency requirements.

+++ src/RecipeVault.DomainService/IngredientSearchService.cs
@@ -197,7 +197,7 @@ namespace RecipeVault.DomainService {
                         TagResourceId = rt.Tag?.TagResourceId ?? Guid.Empty,
                         GlobalName = rt.Tag?.Name,
                         DisplayName = rt.Tag?.Name,
-                        Category = rt.Tag?.Category ?? 0,
+                        Category = (int)(rt.Tag?.Category ?? 0),
                         IsAiAssigned = rt.IsAiAssigned,
                         Confidence = rt.Confidence
                     }).ToList() ?? new List<RecipeTagDto>(),
